minor changes and restructure
