+++ bin/.gitignore
@@ -3,4 +3,3 @@
 .project
 .settings
 .vscode
-/renderer/
