+++ src/renderer/shapes/MyPolygons.java
@@ -9,7 +9,7 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Comparator;
 
-import renderer.input.myVector;
+import renderer.point.myVector;
 import renderer.point.MyPoint;
 import renderer.point.PointConverter;
